requirements.txt
allennlp==0.7.0
networkx>=2.0
pystemmer
cython
pybloof
dill
mmh3
lmdb
